Made review changes mentioned
