+++ RangeColumn_Chart/App.xaml
