+++ RangeColumn_Chart/App.xaml.cs
