+++ RangeColumn_Chart/AssemblyInfo.cs
