+++ RangeColumn_Chart/MainWindow.xaml.cs
